requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Detection
pycocotools

# Re-ID
torchreid
gdown

# Tracking
lapx>=0.5.0
scipy>=1.10.0

# Video / Image
opencv-python>=4.7.0

# Config
PyYAML>=6.0

# API server (optional — only needed for --serve mode)
fastapi
uvicorn[standard]
python-multipart
